+++ ecommerce_project/myapp/serializers/__init__.py
@@ -1,2 +1,3 @@
 from .product_serializers import *
 from .review_serializers import *
+from .cart_serializer import *
